requirements.txt
mysql-connector-python==9.2.0
sql_metadata==2.15.0
sqlparse==0.5.3
